Simple renameing of global vars/defs file name
